% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Load Data}
\usage{
get_data(file)
}
\arguments{
\item{file}{The path to a file you wish to load. This file can be an xlsx, sav, or csv file.}
}
\value{
A data.frame
}
\description{
A function to load data from multiple different file types.
}
\examples{
\dontrun{
dat <- get_data("testfile.sav")
}
}
