% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrRight.R
\name{substrRight}
\alias{substrRight}
\title{Last Letters}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{A character string.}

\item{n}{A number. This is the number of letters from the end of a string to be displayed.}
}
\value{
The last 'n' characters in string 'x'
}
\description{
A function to grab the last 'n' letters from string 'x'
}
\examples{
substrRight("Random String", 3)


}
